Rename materials and definitions
